pyproject.toml
[project]
name = "education"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.3",
    "dj-rest-auth>=7.0.1",
    "django>=5.2",
    "django-allauth>=65.7.0",
    "django-redis>=5.4.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.4.1",
    "firebase-admin>=6.8.0",
    "gunicorn>=23.0.0",
    "locust>=2.36.2",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.7",
    "whitenoise>=6.9.0",
]
